fix: post-review hardening — double-push UAF guard + 6 correctness/doc/CI fixes
Findings from a cross-phase review of the 0.3.0 arc (each phase was reviewed
in isolation; this pass looked at their interaction as a whole).

1. CRITICAL — double-push use-after-free in push_large_deferred_free
   (registry/heap_core.rs). A double-free of the same Large segment from a
   remote thread (UB by the GlobalAlloc contract, but the allocator degrades
   safely on it everywhere else via the M2 guard) could push one `base` onto
   the deferred-free Treiber stack twice before a drain: the second push wrote
   base.next_abandoned = base (self-loop), and a drain then popped+reclaimed
   (unmapped) base while head still pointed at it → next iteration read
   next_abandoned off unmapped memory (UAF) and could double-unmap. A1 had
   silently weakened the pre-existing M2 "double-free is a safe no-op" property
   for exactly the remote-free-Large path.
   Fix: an idempotency guard — a compare_exchange on base's next_abandoned link
   word from ABANDONED_TAIL ("not on any stack") to the encoded link, gating
   entry into the head-CAS loop; a second push of the same base sees the word
   already claimed and no-ops. Required a dedicated on-this-stack tail sentinel
   DEFERRED_LARGE_TAIL (u64::MAX-1) distinct from ABANDONED_TAIL: reusing
   ABANDONED_TAIL as the empty-stack link left next_abandoned == ABANDONED_TAIL
   after a push onto an empty stack, indistinguishable from "never pushed" —
   defeating the guard on first use (the agent caught this via a minimal repro
   during implementation). Multi-producer lock-free push for DISTINCT bases is
   preserved (the guard touches only the pushed segment's own link word).
   Regression test xthread_large_double_free_no_double_reclaim: a remote thread
   double-frees every pointer in a 50×512KiB batch; asserts
   DBG_LARGE_XTHREAD_RECLAIMED advances by EXACTLY N (never <, a leak; never >,
   a double-reclaim) and the heap stays usable. Counterfactual re-verified
   personally: reverting to the unconditional store crashes with a
   deterministic STATUS_ACCESS_VIOLATION 3/3 runs; the guard makes it pass 3/3.
   The guard resets correctly on segment reuse — SegmentHeader::large() (both
   the fresh-OS and large-cache-hit reissue sites) always rewrites
   next_abandoned = ABANDONED_TAIL, so a reused segment is never stuck "claimed".

2. AllocStats::ring_overflows doc was false ("falls back to Treiber-stack, no
   memory is lost"). A ring overflow DISCARDS the block (a bounded leak, per
   remote_free_ring's own "Overflow semantics"). Corrected the public-API doc
   to say so; grep confirmed the phrase appeared nowhere else.

3. perf-gate.yml benched --features alloc-global — WITHOUT fastbin/alloc-decommit,
   i.e. around the very magazine (C1) and large-cache (D1) paths the gate exists
   to guard. Changed to --features production.

4. AllocStats::heaps_claimed_high_water u32 → u64 (the lone non-u64 field;
   widened now, before 0.3.0 publish makes it a breaking change). Cast at the
   stats() call site.

5. run_tsan.sh removed (stale 0.2.1-era session helper with a hardcoded WSL
   path, superseded by scripts/tsan.mjs).

6. size_classes.rs doc: stale SMALL_CLASS_COUNT "(40)" → "(currently 48)" (two
   spots) after B1 grew the table.

7. perf-gate.yml: cargo install iai-callgrind-runner --version 0.14 →
   "^0.14" (0.14 risks being read as an exact, nonexistent version; caret
   matches Cargo.toml).

Verification (personally re-run): production suite --no-fail-fast x2 clean;
loom (8 files / 24 tests) green; the double-free counterfactual crash
reproduced 3/3 and the fix confirmed 3/3; fmt clean; clippy -D warnings 0;
perf-gate.yml valid YAML; cargo build --benches --features production clean.

Fixes implemented by an sh sub-agent; reviewed line-by-line (the guard's
atomic ordering + sentinel distinctness + reset-on-reuse independently
traced), counterfactual independently reproduced before commit.

+++ src/global/alloc_stats.rs
@@ -71,10 +71,15 @@ pub struct AllocStats {
     pub tcache_hits: u64,
 
     /// Number of times a cross-thread free could not be pushed onto a
-    /// segment's remote-free ring because the ring was full (the free then
-    /// falls back to the slower Treiber-stack path — no memory is lost, but
-    /// a high rate here indicates ring-capacity pressure worth tuning).
-    /// Requires the `alloc-xthread` feature; `0` otherwise.
+    /// segment's remote-free ring because the ring was full. On overflow the
+    /// freed block is **discarded** (it stays mapped and unused — a bounded
+    /// leak; see "Overflow semantics" in
+    /// [`remote_free_ring`](crate::alloc_core::remote_free_ring)'s module
+    /// docs). This is sound (no UAF, no corruption) but is NOT free — a
+    /// sustained high rate here means blocks are actually being leaked and
+    /// indicates ring-capacity pressure worth tuning (e.g. a larger ring, or
+    /// more frequent owner-side drains). Requires the `alloc-xthread`
+    /// feature; `0` otherwise.
     pub ring_overflows: u64,
 
     /// Cumulative count of successful OS segment reservations since process
@@ -100,5 +105,9 @@ pub struct AllocStats {
     /// process has needed simultaneously-or-sequentially. This is **not** a
     /// live count: a claimed-then-recycled slot is still counted (recycled
     /// slots are reused for new threads, never un-minted). Always available.
-    pub heaps_claimed_high_water: u32,
+    ///
+    /// `u64` for consistency with every other field on this struct (the
+    /// underlying registry counter is a `u32`; widened here via `as u64` —
+    /// see `SeferAlloc::stats()`).
+    pub heaps_claimed_high_water: u64,
 }
